Moving Allocator Options to seprate file
